requirements.txt
Django==3.2.10
djangorestframework==3.12.4
psycopg2-binary==2.9.1
celery==5.1.2
redis==3.5.3